requirements.txt
Flask
Flask-Login
firebase-admin
Werkzeug
Flask-WTF
python-dotenv
authlib
google-generativeai
requests
gunicorn
email_validator
markdown